Fixed [FMDatabase openWithFlags:] so it does not reopen the db unnecessarily
This is just copying what was done in [FMDatabase open:] by return YES if the database already exists. This solves a problem with issue #211 where any connection changes being made in the [FMDatabasePoolDelegeate databasePool: didAddDatabase:] would just get reset as the database would be reopened on each pool call. Likely inefficent as well.