Added in support for userVersion property
This uses '#PRAGMA user_version' to store/retrieve a uint32 value. According the sqlite3 docs can be used by applications for any purpose. It is useful for tracking database versions, and is used on by the Android SDK calls for that purpose, which can be helpful when doing cross platform work.