Add a copy headers build phase to FMDB-IOS target
Adds a copy headers build phase to the FMDB-IOS target, which
includes the same Public headers as the FMDB target.  Also makes
FMDatabasePool.h a Private (rather than a Project) header in both
FMDB and FMDB-IOS, so if an application imports FMDatabase.h, then
FMDatabase will be able to find FMDatabasePool.h.

This makes it easier to incorporate FMDB-IOS into an iOS
application project.  FMDB can be included in the app's project
as a subproject, and the app target can depend on and link against
FMDB-IOS, and find the FMDB headers in a natural place:
$(BUILT_PRODUCTS_DIR)/usr/local/include.