Move testApplicationID
This is unrelated to the other changes associated with this pull request, but there is a test case which is testing `applicationID` which is was part of core `FMDatabase` tests, and it really should be part of the `FMDatabaseAdditions` test.

It appears that `applicationID` and `applicationIDString` were part of `FMDatabase` and were later moved to `FMDatabaseAdditions`, but that the tests were never updated correspondingly. Fixed.