Return `nil` if the column is NULL
Note, the previous incarnation of `valueForQuery` would return `nil` if the query failed or returned no rows at all, but then return an object if the query succeeded, regardless if the column was `NULL` or not. That didn't make any sense.

So, change this so it returns `nil` if either (a) query failed; or (b) column was `NULL`. Otherwise, rely upon SQLite function to return value of appropriate type.

Note, can still return strange values if the value in the table was the wrong type (e.g. `intForQuery` where the returned column contained a string, e.g. "foo", will return `0` rather than `nil`). But this is a "feature" of SQLite, that dynamically converts between the data type of the column and the function (notably converting from strings to numerics transparently). We would *not* want to actually use the fact that the actual type of the column matched, because there's a long history of SQLite programmers blithely storing numeric values as strings, and we wouldn't want to break that behavior.