Replace deprecated unicode61 dependency and rephrase comments
The unicode61 tokenizer is now enabled by default and the corresponding build
option is no longer available. FMDB/standaone should depend on sqlite3/fts.
I also updated the comment to emphasize that the most recent stable version of
SQLite is built if the standalone subspec is selected, not a custom prebuilt
version.

Btw.: I am the sqlite3 podspec maintainer and am trying to keep the sqlite3 pod
as close to the official stable release schedule as possible: Delay < a few days

Signed-off-by: Clemens Gruber <clemensgru@gmail.com>